Скорректирован список эндпоинтов. Изменён менеджер по умолчанию. Для загрузки из файла добавлена возможность загружать задачи без времени начала и продолжительности
