+++ src/enums/Endpoint.java
@@ -12,12 +12,12 @@ public enum Endpoint {
     DELETE_SUBTASK,
     DELETE_SUBTASKS,
     GET_EPIC,
-    GET_EPICSBUTASKS,
+    GET_EPIC_SUBTASKS,
     GET_EPICS,
     POST_EPIC,
     DELETE_EPIC,
     DELETE_EPICS,
     GET_HISTORY,
-    GET_PRIORITIZEDTASKS,
+    GET_PRIORITIZED_TASKS,
     UNKNOWN
 }
