+++ src/managers/Managers.java
@@ -3,10 +3,12 @@ package managers;
 import history.HistoryManager;
 import history.InMemoryHistoryManager;
 
+import java.io.File;
+
 public class Managers {
     public static TaskManager getDefault() {
-        //return new FileBackedTaskManager(new File(System.getProperty("user.home"), "testFile.txt"));
-        return new InMemoryTaskManager();
+        return FileBackedTaskManager.loadFromFile(new File(System.getProperty("user.home"), "testFile.txt"));
+        //return new InMemoryTaskManager();
     }
 
     public static HistoryManager getDefaultHistory() {
